+++ mne/__init__.py
@@ -11,7 +11,7 @@ from .source_estimate import read_stc, write_stc, SourceEstimate, morph_data, \
 from .surface import read_bem_surfaces, read_surface
 from .source_space import read_source_spaces
 from .epochs import Epochs
-from .label import label_time_courses, read_label
+from .label import label_time_courses, read_label, label_sign_flip
 from .misc import parse_config, read_reject_parameters
 from .transforms import transform_coordinates
 from .proj import read_proj